+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/service/contract/EventPayload.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class EventPayload implements Serializable {
-    private  Long eventId;
+    private Integer eventId;
     private String name;
     private  String description;
     private LocalDateTime startedDateTime;
